Centralize year/qtr/month offset tests
