+++ pandas/tests/tseries/offsets/test_offsets.py
@@ -2749,213 +2749,6 @@ class TestSemiMonthBegin(Base):
         assert_func(result2, exp)
 
 
-class TestBQuarterBegin(Base):
-    _offset = BQuarterBegin
-
-    def test_repr(self):
-        assert (repr(BQuarterBegin()) ==
-                "<BusinessQuarterBegin: startingMonth=3>")
-        assert (repr(BQuarterBegin(startingMonth=3)) ==
-                "<BusinessQuarterBegin: startingMonth=3>")
-        assert (repr(BQuarterBegin(startingMonth=1)) ==
-                "<BusinessQuarterBegin: startingMonth=1>")
-
-    def test_isAnchored(self):
-        assert BQuarterBegin(startingMonth=1).isAnchored()
-        assert BQuarterBegin().isAnchored()
-        assert not BQuarterBegin(2, startingMonth=1).isAnchored()
-
-    offset_cases = []
-    offset_cases.append((BQuarterBegin(startingMonth=1), {
-        datetime(2008, 1, 1): datetime(2008, 4, 1),
-        datetime(2008, 1, 31): datetime(2008, 4, 1),
-        datetime(2008, 2, 15): datetime(2008, 4, 1),
-        datetime(2008, 2, 29): datetime(2008, 4, 1),
-        datetime(2008, 3, 15): datetime(2008, 4, 1),
-        datetime(2008, 3, 31): datetime(2008, 4, 1),
-        datetime(2008, 4, 15): datetime(2008, 7, 1),
-        datetime(2007, 3, 15): datetime(2007, 4, 2),
-        datetime(2007, 2, 28): datetime(2007, 4, 2),
-        datetime(2007, 1, 1): datetime(2007, 4, 2),
-        datetime(2007, 4, 15): datetime(2007, 7, 2),
-        datetime(2007, 7, 1): datetime(2007, 7, 2),
-        datetime(2007, 4, 1): datetime(2007, 4, 2),
-        datetime(2007, 4, 2): datetime(2007, 7, 2),
-        datetime(2008, 4, 30): datetime(2008, 7, 1)}))
-
-    offset_cases.append((BQuarterBegin(startingMonth=2), {
-        datetime(2008, 1, 1): datetime(2008, 2, 1),
-        datetime(2008, 1, 31): datetime(2008, 2, 1),
-        datetime(2008, 1, 15): datetime(2008, 2, 1),
-        datetime(2008, 2, 29): datetime(2008, 5, 1),
-        datetime(2008, 3, 15): datetime(2008, 5, 1),
-        datetime(2008, 3, 31): datetime(2008, 5, 1),
-        datetime(2008, 4, 15): datetime(2008, 5, 1),
-        datetime(2008, 8, 15): datetime(2008, 11, 3),
-        datetime(2008, 9, 15): datetime(2008, 11, 3),
-        datetime(2008, 11, 1): datetime(2008, 11, 3),
-        datetime(2008, 4, 30): datetime(2008, 5, 1)}))
-
-    offset_cases.append((BQuarterBegin(startingMonth=1, n=0), {
-        datetime(2008, 1, 1): datetime(2008, 1, 1),
-        datetime(2007, 12, 31): datetime(2008, 1, 1),
-        datetime(2008, 2, 15): datetime(2008, 4, 1),
-        datetime(2008, 2, 29): datetime(2008, 4, 1),
-        datetime(2008, 1, 15): datetime(2008, 4, 1),
-        datetime(2008, 2, 27): datetime(2008, 4, 1),
-        datetime(2008, 3, 15): datetime(2008, 4, 1),
-        datetime(2007, 4, 1): datetime(2007, 4, 2),
-        datetime(2007, 4, 2): datetime(2007, 4, 2),
-        datetime(2007, 7, 1): datetime(2007, 7, 2),
-        datetime(2007, 4, 15): datetime(2007, 7, 2),
-        datetime(2007, 7, 2): datetime(2007, 7, 2)}))
-
-    offset_cases.append((BQuarterBegin(startingMonth=1, n=-1), {
-        datetime(2008, 1, 1): datetime(2007, 10, 1),
-        datetime(2008, 1, 31): datetime(2008, 1, 1),
-        datetime(2008, 2, 15): datetime(2008, 1, 1),
-        datetime(2008, 2, 29): datetime(2008, 1, 1),
-        datetime(2008, 3, 15): datetime(2008, 1, 1),
-        datetime(2008, 3, 31): datetime(2008, 1, 1),
-        datetime(2008, 4, 15): datetime(2008, 4, 1),
-        datetime(2007, 7, 3): datetime(2007, 7, 2),
-        datetime(2007, 4, 3): datetime(2007, 4, 2),
-        datetime(2007, 7, 2): datetime(2007, 4, 2),
-        datetime(2008, 4, 1): datetime(2008, 1, 1)}))
-
-    offset_cases.append((BQuarterBegin(startingMonth=1, n=2), {
-        datetime(2008, 1, 1): datetime(2008, 7, 1),
-        datetime(2008, 1, 15): datetime(2008, 7, 1),
-        datetime(2008, 2, 29): datetime(2008, 7, 1),
-        datetime(2008, 3, 15): datetime(2008, 7, 1),
-        datetime(2007, 3, 31): datetime(2007, 7, 2),
-        datetime(2007, 4, 15): datetime(2007, 10, 1),
-        datetime(2008, 4, 30): datetime(2008, 10, 1)}))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    def test_offset_corner_case(self):
-        # corner
-        offset = BQuarterBegin(n=-1, startingMonth=1)
-        assert datetime(2007, 4, 3) + offset == datetime(2007, 4, 2)
-
-
-class TestBQuarterEnd(Base):
-    _offset = BQuarterEnd
-
-    def test_repr(self):
-        assert (repr(BQuarterEnd()) ==
-                "<BusinessQuarterEnd: startingMonth=3>")
-        assert (repr(BQuarterEnd(startingMonth=3)) ==
-                "<BusinessQuarterEnd: startingMonth=3>")
-        assert (repr(BQuarterEnd(startingMonth=1)) ==
-                "<BusinessQuarterEnd: startingMonth=1>")
-
-    def test_isAnchored(self):
-        assert BQuarterEnd(startingMonth=1).isAnchored()
-        assert BQuarterEnd().isAnchored()
-        assert not BQuarterEnd(2, startingMonth=1).isAnchored()
-
-    offset_cases = []
-    offset_cases.append((BQuarterEnd(startingMonth=1),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 31),
-                         datetime(2008, 1, 31): datetime(2008, 4, 30),
-                         datetime(2008, 2, 15): datetime(2008, 4, 30),
-                         datetime(2008, 2, 29): datetime(2008, 4, 30),
-                         datetime(2008, 3, 15): datetime(2008, 4, 30),
-                         datetime(2008, 3, 31): datetime(2008, 4, 30),
-                         datetime(2008, 4, 15): datetime(2008, 4, 30),
-                         datetime(2008, 4, 30): datetime(2008, 7, 31), }))
-
-    offset_cases.append((BQuarterEnd(startingMonth=2),
-                        {datetime(2008, 1, 1): datetime(2008, 2, 29),
-                         datetime(2008, 1, 31): datetime(2008, 2, 29),
-                         datetime(2008, 2, 15): datetime(2008, 2, 29),
-                         datetime(2008, 2, 29): datetime(2008, 5, 30),
-                         datetime(2008, 3, 15): datetime(2008, 5, 30),
-                         datetime(2008, 3, 31): datetime(2008, 5, 30),
-                         datetime(2008, 4, 15): datetime(2008, 5, 30),
-                         datetime(2008, 4, 30): datetime(2008, 5, 30), }))
-
-    offset_cases.append((BQuarterEnd(startingMonth=1, n=0),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 31),
-                         datetime(2008, 1, 31): datetime(2008, 1, 31),
-                         datetime(2008, 2, 15): datetime(2008, 4, 30),
-                         datetime(2008, 2, 29): datetime(2008, 4, 30),
-                         datetime(2008, 3, 15): datetime(2008, 4, 30),
-                         datetime(2008, 3, 31): datetime(2008, 4, 30),
-                         datetime(2008, 4, 15): datetime(2008, 4, 30),
-                         datetime(2008, 4, 30): datetime(2008, 4, 30), }))
-
-    offset_cases.append((BQuarterEnd(startingMonth=1, n=-1),
-                        {datetime(2008, 1, 1): datetime(2007, 10, 31),
-                         datetime(2008, 1, 31): datetime(2007, 10, 31),
-                         datetime(2008, 2, 15): datetime(2008, 1, 31),
-                         datetime(2008, 2, 29): datetime(2008, 1, 31),
-                         datetime(2008, 3, 15): datetime(2008, 1, 31),
-                         datetime(2008, 3, 31): datetime(2008, 1, 31),
-                         datetime(2008, 4, 15): datetime(2008, 1, 31),
-                         datetime(2008, 4, 30): datetime(2008, 1, 31), }))
-
-    offset_cases.append((BQuarterEnd(startingMonth=1, n=2),
-                        {datetime(2008, 1, 31): datetime(2008, 7, 31),
-                         datetime(2008, 2, 15): datetime(2008, 7, 31),
-                         datetime(2008, 2, 29): datetime(2008, 7, 31),
-                         datetime(2008, 3, 15): datetime(2008, 7, 31),
-                         datetime(2008, 3, 31): datetime(2008, 7, 31),
-                         datetime(2008, 4, 15): datetime(2008, 7, 31),
-                         datetime(2008, 4, 30): datetime(2008, 10, 31), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    def test_offset_corner_case(self):
-        # corner
-        offset = BQuarterEnd(n=-1, startingMonth=1)
-        assert datetime(2010, 1, 31) + offset == datetime(2010, 1, 29)
-
-    on_offset_cases = [
-        (BQuarterEnd(1, startingMonth=1), datetime(2008, 1, 31), True),
-        (BQuarterEnd(1, startingMonth=1), datetime(2007, 12, 31), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2008, 2, 29), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2007, 3, 30), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2007, 3, 31), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2008, 4, 30), True),
-        (BQuarterEnd(1, startingMonth=1), datetime(2008, 5, 30), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2007, 6, 29), False),
-        (BQuarterEnd(1, startingMonth=1), datetime(2007, 6, 30), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2008, 1, 31), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2007, 12, 31), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2008, 2, 29), True),
-        (BQuarterEnd(1, startingMonth=2), datetime(2007, 3, 30), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2007, 3, 31), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2008, 4, 30), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2008, 5, 30), True),
-        (BQuarterEnd(1, startingMonth=2), datetime(2007, 6, 29), False),
-        (BQuarterEnd(1, startingMonth=2), datetime(2007, 6, 30), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2008, 1, 31), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2007, 12, 31), True),
-        (BQuarterEnd(1, startingMonth=3), datetime(2008, 2, 29), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2007, 3, 30), True),
-        (BQuarterEnd(1, startingMonth=3), datetime(2007, 3, 31), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2008, 4, 30), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2008, 5, 30), False),
-        (BQuarterEnd(1, startingMonth=3), datetime(2007, 6, 29), True),
-        (BQuarterEnd(1, startingMonth=3), datetime(2007, 6, 30), False)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
 def makeFY5253LastOfMonthQuarter(*args, **kwds):
     return FY5253Quarter(*args, variation="last", **kwds)
 
@@ -3506,519 +3299,6 @@ class TestFY5253NearestEndMonthQuarter(Base):
                             datetime(2013, 3, 30))
 
 
-class TestQuarterBegin(Base):
-
-    def test_repr(self):
-        assert (repr(QuarterBegin()) ==
-                "<QuarterBegin: startingMonth=3>")
-        assert (repr(QuarterBegin(startingMonth=3)) ==
-                "<QuarterBegin: startingMonth=3>")
-        assert (repr(QuarterBegin(startingMonth=1)) ==
-                "<QuarterBegin: startingMonth=1>")
-
-    def test_isAnchored(self):
-        assert QuarterBegin(startingMonth=1).isAnchored()
-        assert QuarterBegin().isAnchored()
-        assert not QuarterBegin(2, startingMonth=1).isAnchored()
-
-    offset_cases = []
-    offset_cases.append((QuarterBegin(startingMonth=1),
-                        {datetime(2007, 12, 1): datetime(2008, 1, 1),
-                         datetime(2008, 1, 1): datetime(2008, 4, 1),
-                         datetime(2008, 2, 15): datetime(2008, 4, 1),
-                         datetime(2008, 2, 29): datetime(2008, 4, 1),
-                         datetime(2008, 3, 15): datetime(2008, 4, 1),
-                         datetime(2008, 3, 31): datetime(2008, 4, 1),
-                         datetime(2008, 4, 15): datetime(2008, 7, 1),
-                         datetime(2008, 4, 1): datetime(2008, 7, 1), }))
-
-    offset_cases.append((QuarterBegin(startingMonth=2),
-                        {datetime(2008, 1, 1): datetime(2008, 2, 1),
-                         datetime(2008, 1, 31): datetime(2008, 2, 1),
-                         datetime(2008, 1, 15): datetime(2008, 2, 1),
-                         datetime(2008, 2, 29): datetime(2008, 5, 1),
-                         datetime(2008, 3, 15): datetime(2008, 5, 1),
-                         datetime(2008, 3, 31): datetime(2008, 5, 1),
-                         datetime(2008, 4, 15): datetime(2008, 5, 1),
-                         datetime(2008, 4, 30): datetime(2008, 5, 1), }))
-
-    offset_cases.append((QuarterBegin(startingMonth=1, n=0),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 1),
-                         datetime(2008, 12, 1): datetime(2009, 1, 1),
-                         datetime(2008, 1, 1): datetime(2008, 1, 1),
-                         datetime(2008, 2, 15): datetime(2008, 4, 1),
-                         datetime(2008, 2, 29): datetime(2008, 4, 1),
-                         datetime(2008, 3, 15): datetime(2008, 4, 1),
-                         datetime(2008, 3, 31): datetime(2008, 4, 1),
-                         datetime(2008, 4, 15): datetime(2008, 7, 1),
-                         datetime(2008, 4, 30): datetime(2008, 7, 1), }))
-
-    offset_cases.append((QuarterBegin(startingMonth=1, n=-1),
-                        {datetime(2008, 1, 1): datetime(2007, 10, 1),
-                         datetime(2008, 1, 31): datetime(2008, 1, 1),
-                         datetime(2008, 2, 15): datetime(2008, 1, 1),
-                         datetime(2008, 2, 29): datetime(2008, 1, 1),
-                         datetime(2008, 3, 15): datetime(2008, 1, 1),
-                         datetime(2008, 3, 31): datetime(2008, 1, 1),
-                         datetime(2008, 4, 15): datetime(2008, 4, 1),
-                         datetime(2008, 4, 30): datetime(2008, 4, 1),
-                         datetime(2008, 7, 1): datetime(2008, 4, 1)}))
-
-    offset_cases.append((QuarterBegin(startingMonth=1, n=2),
-                        {datetime(2008, 1, 1): datetime(2008, 7, 1),
-                         datetime(2008, 2, 15): datetime(2008, 7, 1),
-                         datetime(2008, 2, 29): datetime(2008, 7, 1),
-                         datetime(2008, 3, 15): datetime(2008, 7, 1),
-                         datetime(2008, 3, 31): datetime(2008, 7, 1),
-                         datetime(2008, 4, 15): datetime(2008, 10, 1),
-                         datetime(2008, 4, 1): datetime(2008, 10, 1), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    def test_offset_corner_case(self):
-        # corner
-        offset = QuarterBegin(n=-1, startingMonth=1)
-        assert datetime(2010, 2, 1) + offset == datetime(2010, 1, 1)
-
-
-class TestQuarterEnd(Base):
-    _offset = QuarterEnd
-
-    def test_repr(self):
-        assert (repr(QuarterEnd()) ==
-                "<QuarterEnd: startingMonth=3>")
-        assert (repr(QuarterEnd(startingMonth=3)) ==
-                "<QuarterEnd: startingMonth=3>")
-        assert (repr(QuarterEnd(startingMonth=1)) ==
-                "<QuarterEnd: startingMonth=1>")
-
-    def test_isAnchored(self):
-        assert QuarterEnd(startingMonth=1).isAnchored()
-        assert QuarterEnd().isAnchored()
-        assert not QuarterEnd(2, startingMonth=1).isAnchored()
-
-    offset_cases = []
-    offset_cases.append((QuarterEnd(startingMonth=1),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 31),
-                         datetime(2008, 1, 31): datetime(2008, 4, 30),
-                         datetime(2008, 2, 15): datetime(2008, 4, 30),
-                         datetime(2008, 2, 29): datetime(2008, 4, 30),
-                         datetime(2008, 3, 15): datetime(2008, 4, 30),
-                         datetime(2008, 3, 31): datetime(2008, 4, 30),
-                         datetime(2008, 4, 15): datetime(2008, 4, 30),
-                         datetime(2008, 4, 30): datetime(2008, 7, 31), }))
-
-    offset_cases.append((QuarterEnd(startingMonth=2),
-                        {datetime(2008, 1, 1): datetime(2008, 2, 29),
-                         datetime(2008, 1, 31): datetime(2008, 2, 29),
-                         datetime(2008, 2, 15): datetime(2008, 2, 29),
-                         datetime(2008, 2, 29): datetime(2008, 5, 31),
-                         datetime(2008, 3, 15): datetime(2008, 5, 31),
-                         datetime(2008, 3, 31): datetime(2008, 5, 31),
-                         datetime(2008, 4, 15): datetime(2008, 5, 31),
-                         datetime(2008, 4, 30): datetime(2008, 5, 31), }))
-
-    offset_cases.append((QuarterEnd(startingMonth=1, n=0),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 31),
-                         datetime(2008, 1, 31): datetime(2008, 1, 31),
-                         datetime(2008, 2, 15): datetime(2008, 4, 30),
-                         datetime(2008, 2, 29): datetime(2008, 4, 30),
-                         datetime(2008, 3, 15): datetime(2008, 4, 30),
-                         datetime(2008, 3, 31): datetime(2008, 4, 30),
-                         datetime(2008, 4, 15): datetime(2008, 4, 30),
-                         datetime(2008, 4, 30): datetime(2008, 4, 30), }))
-
-    offset_cases.append((QuarterEnd(startingMonth=1, n=-1),
-                        {datetime(2008, 1, 1): datetime(2007, 10, 31),
-                         datetime(2008, 1, 31): datetime(2007, 10, 31),
-                         datetime(2008, 2, 15): datetime(2008, 1, 31),
-                         datetime(2008, 2, 29): datetime(2008, 1, 31),
-                         datetime(2008, 3, 15): datetime(2008, 1, 31),
-                         datetime(2008, 3, 31): datetime(2008, 1, 31),
-                         datetime(2008, 4, 15): datetime(2008, 1, 31),
-                         datetime(2008, 4, 30): datetime(2008, 1, 31),
-                         datetime(2008, 7, 1): datetime(2008, 4, 30)}))
-
-    offset_cases.append((QuarterEnd(startingMonth=1, n=2),
-                        {datetime(2008, 1, 31): datetime(2008, 7, 31),
-                         datetime(2008, 2, 15): datetime(2008, 7, 31),
-                         datetime(2008, 2, 29): datetime(2008, 7, 31),
-                         datetime(2008, 3, 15): datetime(2008, 7, 31),
-                         datetime(2008, 3, 31): datetime(2008, 7, 31),
-                         datetime(2008, 4, 15): datetime(2008, 7, 31),
-                         datetime(2008, 4, 30): datetime(2008, 10, 31), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    def test_offset_corner_case(self):
-        # corner
-        offset = QuarterEnd(n=-1, startingMonth=1)
-        assert datetime(2010, 2, 1) + offset == datetime(2010, 1, 31)
-
-    on_offset_cases = [
-        (QuarterEnd(1, startingMonth=1), datetime(2008, 1, 31), True),
-        (QuarterEnd(1, startingMonth=1), datetime(2007, 12, 31), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2008, 2, 29), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2007, 3, 30), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2007, 3, 31), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2008, 4, 30), True),
-        (QuarterEnd(1, startingMonth=1), datetime(2008, 5, 30), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2008, 5, 31), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2007, 6, 29), False),
-        (QuarterEnd(1, startingMonth=1), datetime(2007, 6, 30), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2008, 1, 31), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2007, 12, 31), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2008, 2, 29), True),
-        (QuarterEnd(1, startingMonth=2), datetime(2007, 3, 30), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2007, 3, 31), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2008, 4, 30), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2008, 5, 30), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2008, 5, 31), True),
-        (QuarterEnd(1, startingMonth=2), datetime(2007, 6, 29), False),
-        (QuarterEnd(1, startingMonth=2), datetime(2007, 6, 30), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2008, 1, 31), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2007, 12, 31), True),
-        (QuarterEnd(1, startingMonth=3), datetime(2008, 2, 29), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2007, 3, 30), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2007, 3, 31), True),
-        (QuarterEnd(1, startingMonth=3), datetime(2008, 4, 30), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2008, 5, 30), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2008, 5, 31), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2007, 6, 29), False),
-        (QuarterEnd(1, startingMonth=3), datetime(2007, 6, 30), True)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
-class TestBYearBegin(Base):
-    _offset = BYearBegin
-
-    def test_misspecified(self):
-        pytest.raises(ValueError, BYearBegin, month=13)
-        pytest.raises(ValueError, BYearEnd, month=13)
-
-    offset_cases = []
-    offset_cases.append((BYearBegin(),
-                        {datetime(2008, 1, 1): datetime(2009, 1, 1),
-                         datetime(2008, 6, 30): datetime(2009, 1, 1),
-                         datetime(2008, 12, 31): datetime(2009, 1, 1),
-                         datetime(2011, 1, 1): datetime(2011, 1, 3),
-                         datetime(2011, 1, 3): datetime(2012, 1, 2),
-                         datetime(2005, 12, 30): datetime(2006, 1, 2),
-                         datetime(2005, 12, 31): datetime(2006, 1, 2)}))
-
-    offset_cases.append((BYearBegin(0),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 1),
-                         datetime(2008, 6, 30): datetime(2009, 1, 1),
-                         datetime(2008, 12, 31): datetime(2009, 1, 1),
-                         datetime(2005, 12, 30): datetime(2006, 1, 2),
-                         datetime(2005, 12, 31): datetime(2006, 1, 2), }))
-
-    offset_cases.append((BYearBegin(-1),
-                        {datetime(2007, 1, 1): datetime(2006, 1, 2),
-                         datetime(2009, 1, 4): datetime(2009, 1, 1),
-                         datetime(2009, 1, 1): datetime(2008, 1, 1),
-                         datetime(2008, 6, 30): datetime(2008, 1, 1),
-                         datetime(2008, 12, 31): datetime(2008, 1, 1),
-                         datetime(2006, 12, 29): datetime(2006, 1, 2),
-                         datetime(2006, 12, 30): datetime(2006, 1, 2),
-                         datetime(2006, 1, 1): datetime(2005, 1, 3), }))
-
-    offset_cases.append((BYearBegin(-2),
-                        {datetime(2007, 1, 1): datetime(2005, 1, 3),
-                         datetime(2007, 6, 30): datetime(2006, 1, 2),
-                         datetime(2008, 12, 31): datetime(2007, 1, 1), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-
-class TestYearBegin(Base):
-    _offset = YearBegin
-
-    def test_misspecified(self):
-        pytest.raises(ValueError, YearBegin, month=13)
-
-    offset_cases = []
-    offset_cases.append((YearBegin(),
-                        {datetime(2008, 1, 1): datetime(2009, 1, 1),
-                         datetime(2008, 6, 30): datetime(2009, 1, 1),
-                         datetime(2008, 12, 31): datetime(2009, 1, 1),
-                         datetime(2005, 12, 30): datetime(2006, 1, 1),
-                         datetime(2005, 12, 31): datetime(2006, 1, 1), }))
-
-    offset_cases.append((YearBegin(0),
-                        {datetime(2008, 1, 1): datetime(2008, 1, 1),
-                         datetime(2008, 6, 30): datetime(2009, 1, 1),
-                         datetime(2008, 12, 31): datetime(2009, 1, 1),
-                         datetime(2005, 12, 30): datetime(2006, 1, 1),
-                         datetime(2005, 12, 31): datetime(2006, 1, 1), }))
-
-    offset_cases.append((YearBegin(3),
-                        {datetime(2008, 1, 1): datetime(2011, 1, 1),
-                         datetime(2008, 6, 30): datetime(2011, 1, 1),
-                         datetime(2008, 12, 31): datetime(2011, 1, 1),
-                         datetime(2005, 12, 30): datetime(2008, 1, 1),
-                         datetime(2005, 12, 31): datetime(2008, 1, 1), }))
-
-    offset_cases.append((YearBegin(-1),
-                        {datetime(2007, 1, 1): datetime(2006, 1, 1),
-                         datetime(2007, 1, 15): datetime(2007, 1, 1),
-                         datetime(2008, 6, 30): datetime(2008, 1, 1),
-                         datetime(2008, 12, 31): datetime(2008, 1, 1),
-                         datetime(2006, 12, 29): datetime(2006, 1, 1),
-                         datetime(2006, 12, 30): datetime(2006, 1, 1),
-                         datetime(2007, 1, 1): datetime(2006, 1, 1), }))
-
-    offset_cases.append((YearBegin(-2),
-                        {datetime(2007, 1, 1): datetime(2005, 1, 1),
-                         datetime(2008, 6, 30): datetime(2007, 1, 1),
-                         datetime(2008, 12, 31): datetime(2007, 1, 1), }))
-
-    offset_cases.append((YearBegin(month=4),
-                        {datetime(2007, 4, 1): datetime(2008, 4, 1),
-                         datetime(2007, 4, 15): datetime(2008, 4, 1),
-                         datetime(2007, 3, 1): datetime(2007, 4, 1),
-                         datetime(2007, 12, 15): datetime(2008, 4, 1),
-                         datetime(2012, 1, 31): datetime(2012, 4, 1), }))
-
-    offset_cases.append((YearBegin(0, month=4),
-                        {datetime(2007, 4, 1): datetime(2007, 4, 1),
-                         datetime(2007, 3, 1): datetime(2007, 4, 1),
-                         datetime(2007, 12, 15): datetime(2008, 4, 1),
-                         datetime(2012, 1, 31): datetime(2012, 4, 1), }))
-
-    offset_cases.append((YearBegin(4, month=4),
-                        {datetime(2007, 4, 1): datetime(2011, 4, 1),
-                         datetime(2007, 4, 15): datetime(2011, 4, 1),
-                         datetime(2007, 3, 1): datetime(2010, 4, 1),
-                         datetime(2007, 12, 15): datetime(2011, 4, 1),
-                         datetime(2012, 1, 31): datetime(2015, 4, 1), }))
-
-    offset_cases.append((YearBegin(-1, month=4),
-                        {datetime(2007, 4, 1): datetime(2006, 4, 1),
-                         datetime(2007, 3, 1): datetime(2006, 4, 1),
-                         datetime(2007, 12, 15): datetime(2007, 4, 1),
-                         datetime(2012, 1, 31): datetime(2011, 4, 1), }))
-
-    offset_cases.append((YearBegin(-3, month=4),
-                        {datetime(2007, 4, 1): datetime(2004, 4, 1),
-                         datetime(2007, 3, 1): datetime(2004, 4, 1),
-                         datetime(2007, 12, 15): datetime(2005, 4, 1),
-                         datetime(2012, 1, 31): datetime(2009, 4, 1), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [(YearBegin(), datetime(2007, 1, 3), False),
-                       (YearBegin(), datetime(2008, 1, 1), True),
-                       (YearBegin(), datetime(2006, 12, 31), False),
-                       (YearBegin(), datetime(2006, 1, 2), False)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
-class TestBYearEndLagged(Base):
-
-    def test_bad_month_fail(self):
-        pytest.raises(Exception, BYearEnd, month=13)
-        pytest.raises(Exception, BYearEnd, month=0)
-
-    offset_cases = []
-    offset_cases.append((BYearEnd(month=6),
-                         {datetime(2008, 1, 1): datetime(2008, 6, 30),
-                          datetime(2007, 6, 30): datetime(2008, 6, 30)}, ))
-
-    offset_cases.append((BYearEnd(n=-1, month=6),
-                        {datetime(2008, 1, 1): datetime(2007, 6, 29),
-                         datetime(2007, 6, 30): datetime(2007, 6, 29)}, ))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert base + offset == expected
-
-    def test_roll(self):
-        offset = BYearEnd(month=6)
-        date = datetime(2009, 11, 30)
-
-        assert offset.rollforward(date) == datetime(2010, 6, 30)
-        assert offset.rollback(date) == datetime(2009, 6, 30)
-
-    on_offset_cases = [(BYearEnd(month=2), datetime(2007, 2, 28), True),
-                       (BYearEnd(month=6), datetime(2007, 6, 30), False)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
-class TestBYearEnd(Base):
-    _offset = BYearEnd
-
-    offset_cases = []
-    offset_cases.append((BYearEnd(),
-                        {datetime(2008, 1, 1): datetime(2008, 12, 31),
-                         datetime(2008, 6, 30): datetime(2008, 12, 31),
-                         datetime(2008, 12, 31): datetime(2009, 12, 31),
-                         datetime(2005, 12, 30): datetime(2006, 12, 29),
-                         datetime(2005, 12, 31): datetime(2006, 12, 29), }))
-
-    offset_cases.append((BYearEnd(0),
-                        {datetime(2008, 1, 1): datetime(2008, 12, 31),
-                         datetime(2008, 6, 30): datetime(2008, 12, 31),
-                         datetime(2008, 12, 31): datetime(2008, 12, 31),
-                         datetime(2005, 12, 31): datetime(2006, 12, 29), }))
-
-    offset_cases.append((BYearEnd(-1),
-                        {datetime(2007, 1, 1): datetime(2006, 12, 29),
-                         datetime(2008, 6, 30): datetime(2007, 12, 31),
-                         datetime(2008, 12, 31): datetime(2007, 12, 31),
-                         datetime(2006, 12, 29): datetime(2005, 12, 30),
-                         datetime(2006, 12, 30): datetime(2006, 12, 29),
-                         datetime(2007, 1, 1): datetime(2006, 12, 29), }))
-
-    offset_cases.append((BYearEnd(-2),
-                        {datetime(2007, 1, 1): datetime(2005, 12, 30),
-                         datetime(2008, 6, 30): datetime(2006, 12, 29),
-                         datetime(2008, 12, 31): datetime(2006, 12, 29), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [(BYearEnd(), datetime(2007, 12, 31), True),
-                       (BYearEnd(), datetime(2008, 1, 1), False),
-                       (BYearEnd(), datetime(2006, 12, 31), False),
-                       (BYearEnd(), datetime(2006, 12, 29), True)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
-class TestYearEnd(Base):
-    _offset = YearEnd
-
-    def test_misspecified(self):
-        pytest.raises(ValueError, YearEnd, month=13)
-
-    offset_cases = []
-    offset_cases.append((YearEnd(),
-                        {datetime(2008, 1, 1): datetime(2008, 12, 31),
-                         datetime(2008, 6, 30): datetime(2008, 12, 31),
-                         datetime(2008, 12, 31): datetime(2009, 12, 31),
-                         datetime(2005, 12, 30): datetime(2005, 12, 31),
-                         datetime(2005, 12, 31): datetime(2006, 12, 31), }))
-
-    offset_cases.append((YearEnd(0),
-                        {datetime(2008, 1, 1): datetime(2008, 12, 31),
-                         datetime(2008, 6, 30): datetime(2008, 12, 31),
-                         datetime(2008, 12, 31): datetime(2008, 12, 31),
-                         datetime(2005, 12, 30): datetime(2005, 12, 31), }))
-
-    offset_cases.append((YearEnd(-1),
-                        {datetime(2007, 1, 1): datetime(2006, 12, 31),
-                         datetime(2008, 6, 30): datetime(2007, 12, 31),
-                         datetime(2008, 12, 31): datetime(2007, 12, 31),
-                         datetime(2006, 12, 29): datetime(2005, 12, 31),
-                         datetime(2006, 12, 30): datetime(2005, 12, 31),
-                         datetime(2007, 1, 1): datetime(2006, 12, 31), }))
-
-    offset_cases.append((YearEnd(-2),
-                        {datetime(2007, 1, 1): datetime(2005, 12, 31),
-                         datetime(2008, 6, 30): datetime(2006, 12, 31),
-                         datetime(2008, 12, 31): datetime(2006, 12, 31), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [(YearEnd(), datetime(2007, 12, 31), True),
-                       (YearEnd(), datetime(2008, 1, 1), False),
-                       (YearEnd(), datetime(2006, 12, 31), True),
-                       (YearEnd(), datetime(2006, 12, 29), False)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
-class TestYearEndDiffMonth(Base):
-
-    offset_cases = []
-    offset_cases.append((YearEnd(month=3),
-                        {datetime(2008, 1, 1): datetime(2008, 3, 31),
-                         datetime(2008, 2, 15): datetime(2008, 3, 31),
-                         datetime(2008, 3, 31): datetime(2009, 3, 31),
-                         datetime(2008, 3, 30): datetime(2008, 3, 31),
-                         datetime(2005, 3, 31): datetime(2006, 3, 31),
-                         datetime(2006, 7, 30): datetime(2007, 3, 31)}))
-
-    offset_cases.append((YearEnd(0, month=3),
-                        {datetime(2008, 1, 1): datetime(2008, 3, 31),
-                         datetime(2008, 2, 28): datetime(2008, 3, 31),
-                         datetime(2008, 3, 31): datetime(2008, 3, 31),
-                         datetime(2005, 3, 30): datetime(2005, 3, 31), }))
-
-    offset_cases.append((YearEnd(-1, month=3),
-                        {datetime(2007, 1, 1): datetime(2006, 3, 31),
-                         datetime(2008, 2, 28): datetime(2007, 3, 31),
-                         datetime(2008, 3, 31): datetime(2007, 3, 31),
-                         datetime(2006, 3, 29): datetime(2005, 3, 31),
-                         datetime(2006, 3, 30): datetime(2005, 3, 31),
-                         datetime(2007, 3, 1): datetime(2006, 3, 31), }))
-
-    offset_cases.append((YearEnd(-2, month=3),
-                        {datetime(2007, 1, 1): datetime(2005, 3, 31),
-                         datetime(2008, 6, 30): datetime(2007, 3, 31),
-                         datetime(2008, 3, 31): datetime(2006, 3, 31), }))
-
-    @pytest.mark.parametrize('case', offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in compat.iteritems(cases):
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [(YearEnd(month=3), datetime(2007, 3, 31), True),
-                       (YearEnd(month=3), datetime(2008, 1, 1), False),
-                       (YearEnd(month=3), datetime(2006, 3, 31), True),
-                       (YearEnd(month=3), datetime(2006, 3, 29), False)]
-
-    @pytest.mark.parametrize('case', on_offset_cases)
-    def test_onOffset(self, case):
-        offset, dt, expected = case
-        assert_onOffset(offset, dt, expected)
-
-
 def test_Easter():
     assert_offset_equal(Easter(), datetime(2010, 1, 1), datetime(2010, 4, 4))
     assert_offset_equal(Easter(), datetime(2010, 4, 5), datetime(2011, 4, 24))
@@ -4146,16 +3426,6 @@ def test_get_standard_freq():
 
     with tm.assert_produces_warning(FutureWarning, check_stacklevel=False):
         assert fstr == get_standard_freq(('q', 5))
-
-
-def test_quarterly_dont_normalize():
-    date = datetime(2012, 3, 31, 5, 30)
-
-    offsets = (QuarterBegin, QuarterEnd, BQuarterEnd, BQuarterBegin)
-
-    for klass in offsets:
-        result = date + klass()
-        assert (result.time() == date.time())
 
 
 class TestOffsetAliases(object):
